Slew of Misc Changes
A bunch of fixes, changes, and reworks that I won't be able to fully describe here. Created a registry and an element viewer, to replace the old entity based one. Rewrote how entity lists are stored, and improved upon those. Added a few new conversion and helper scripts to ease fetching of things. Things may be broken, but fixing them should be much easier with these tools